Treeview class to own file
